alteração migration id companies

+++ sistem_orcamento/database/migrations/2025_08_22_161537_add_company_id_to_users_table.php
@@ -11,10 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('company_id')->nullable()->after('role');
-            $table->foreign('company_id')->references('id')->on('companies')->onDelete('set null');
-        });
+        // Verifica se a tabela e a coluna existem para evitar erros
+        if (Schema::hasTable('users') && Schema::hasColumn('users', 'company_id')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->foreign('company_id')->references('id')->on('companies')->onDelete('set null');
+            });
+        }
     }
 
     /**
@@ -24,7 +26,6 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             $table->dropForeign(['company_id']);
-            $table->dropColumn('company_id');
         });
     }
 };
